Move {database,service}_classes => {database,services}
