+++ Dockerfile
@@ -5,10 +5,9 @@ COPY requirements.txt /
 RUN pip install -r /requirements.txt
 
 COPY cogs/ /app/cogs/
-COPY database_classes/ /app/database_classes/
+COPY database/ /app/database/
+COPY services/ /app/services/
 COPY main.py /app
-
-COPY service_classes/ /app/service_classes/
 
 WORKDIR /app
