+++ cogs/fun-commands.py
@@ -50,27 +50,6 @@ class FunCommands(commands.Cog, name="Fun Commands"):
         if ctx.channel.id == self.random_channel:
             await ctx.send("wow that's a long cheat code")
 
-    @commands.command()
-    async def pledge(self, ctx):
-        if ctx.message.author.voice is None:
-            await ctx.send("Please join a voice channel")
-        else:
-            voice_channel = ctx.message.author.voice.channel
-            channel = None
-            if voice_channel != None:
-                channel = voice_channel
-                vc = await channel.connect()
-                os.chdir("./audiofiles")
-                people = []
-                for file in glob.glob("*.mp3"):
-                    people.append(file)
-                person = choice(people)
-                source = discord.FFmpegPCMAudio(f'{person}')
-                player = vc.play(source)
-                while vc.is_playing():
-                    await asyncio.sleep(1)
-                await vc.disconnect()
-
 
 def setup(bot):
     bot.add_cog(FunCommands(bot))
